requirements.txt
ipaddress>=1.0.23
numpy>=1.24.2
openpyxl>=3.0.9
python-pptx>=0.6.21
tkinterdnd2>=0.3.0